Empty chronograp/__init__.py and put in utils.py, removing stale functions
