+++ kalite/updates/api_views.py
@@ -20,7 +20,7 @@ from django.utils.translation import ugettext as _
 from . import delete_downloaded_files, get_local_video_size, get_remote_video_size
 from .models import UpdateProgressLog, VideoFile
 from .views import get_installed_language_packs
-from fle_utils.chronograph import force_job
+from fle_utils.chronograph.utils import force_job
 from fle_utils.django_utils.command import call_command_async
 from fle_utils.general import isnumeric, break_into_chunks, softload_json
 from fle_utils.internet.decorators import api_handle_error_with_json
