+++ python-packages/securesync/devices/views.py
@@ -20,7 +20,7 @@ from .forms import RegisteredDevicePublicKeyForm
 from .models import Device, Zone, RegisteredDevicePublicKey
 from .. import crypto
 from ..engine.models import SyncSession
-from fle_utils.chronograph import force_job
+from fle_utils.chronograph.utils import force_job
 from fle_utils.config.models import Settings
 from fle_utils.internet.classes import JsonResponse
 from fle_utils.internet.decorators import allow_jsonp
