+++ python-packages/securesync/engine/api_views.py
@@ -17,7 +17,7 @@ from django.views.decorators.gzip import gzip_page
 from . import get_serialized_models, save_serialized_models, get_device_counters, serialize
 from .models import *
 from ..devices.models import *  # inter-dependence
-from fle_utils.chronograph import force_job
+from fle_utils.chronograph.utils import force_job
 from fle_utils.django_utils.functions import get_request_ip
 from fle_utils.internet.decorators import api_handle_error_with_json
 from fle_utils.internet.classes import JsonResponse, JsonResponseMessageError
